requirements.txt
aiofiles~=0.8.0
alembic~=1.7.7
aniso8601~=9.0.1
anyio~=3.5.0
asgiref~=3.5.0
async-exit-stack~=1.0.1
async-generator~=1.10
bcrypt~=3.2.0
certifi~=2021.10.8
cffi~=1.15.0
charset-normalizer~=2.0.12
click~=8.0.4
cryptography~=36.0.2
dnspython~=2.2.1
ecdsa~=0.17.0
email-validator~=1.1.3
fastapi~=0.58.1
graphene~=3.0
graphql-core~=3.1.7
graphql-relay~=3.1.5
greenlet~=1.1.2
h11~=0.13.0
idna~=3.3
itsdangerous~=2.1.2
Jinja2~=3.1.0
Mako~=1.2.0
MarkupSafe~=2.1.1
orjson~=3.6.7
passlib~=1.7.4
psycopg2-binary~=2.9.3
pyasn1~=0.4.8
pycparser~=2.21
pydantic~=1.9.0
python-dotenv~=0.20.0
python-jose~=3.3.0
python-multipart~=0.0.4
PyYAML~=5.4.1
requests~=2.27.1
rsa~=4.8
six~=1.16.0
sniffio~=1.2.0
SQLAlchemy~=1.4.32
starlette~=0.13.4
typing_extensions~=4.1.1
ujson~=5.1.0
urllib3~=1.26.9
uvicorn~=0.15.0
uvloop~=0.16.0
watchgod~=0.8.1
websockets~=10.2
pytest~=7.1.1
